pyproject.toml
[tool.poetry]
name = "homework-bot"
version = "0.1.0"
description = ""
authors = ["Ежов Валерий <ezhov.fzzz@gmail.com>"]
readme = "README.md"
packages = [{include = "homework_bot"}]

[tool.poetry.dependencies]
python = "^3.7"

[tool.poetry.group.dev.dependencies]
flake8 = "3.9.2"
flake8-docstrings = "1.6.0"
pytest = "6.2.5"
python-dotenv = "0.19.0"
python-telegram-bot = "13.7"
requests = "2.26.0"
flake8-absolute-import = "^1.0.0.1"
flake8-bugbear = "^22.10.27"
flake8-broken-line = "^0.6.0"
flake8-cognitive-complexity = "^0.1.0"
flake8-commas = "^2.1.0"
flake8-comprehensions = "^3.10.1"
flake8-eradicate = "^1.4.0"
flake8-fixme = "^1.1.1"
flake8-isort = "^5.0.3"
flake8-multiline-containers = "^0.0.19"
flake8-mutable = "^1.2.0"
flake8-pep3101 = "^2.0.0"
flake8-pie = "^0.16.0"
flake8-print = "^5.0.0"
flake8-printf-formatting = "^1.1.2"
flake8-quotes = "^3.3.1"
flake8-return = "^1.2.0"
flake8-simplify = "^0.19.3"
flake8-todo = "^0.7"
pep8-naming = "^0.13.2"
flake8-variables-names = "^0.0.5"
flake8-use-fstring = "^1.4"
black = "^22.10.0"
importlib-metadata = "4.13.0"
darglint = "^1.8.1"
flake8-logging-format = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
